+++ searchguard/__init__.py
@@ -1,0 +1,2 @@
+from .internalusers import *
+from .roles import *
