+++ tests/tests_internalusers/test_delete_user.py
@@ -3,7 +3,7 @@
 import unittest.mock as mock
 from unittest.mock import Mock
 from tests.helper import BaseTestCase
-from searchguard.searchguard import delete_user
+from searchguard.internalusers import delete_user
 from searchguard.exceptions import DeleteUserException
 
 
@@ -12,11 +12,11 @@ class TestDeleteUser(BaseTestCase):
     def setUp(self):
         self.user = "DummyUser"
         self.api_url = "fake_api_url/internalusers/"
-        self.set_up_patch('searchguard.searchguard.SGAPI', "fake_api_url")
+        self.set_up_patch('searchguard.internalusers.SGAPI', "fake_api_url")
 
-        self.mocked_requests_delete = self.set_up_patch('searchguard.searchguard.requests.delete')
+        self.mocked_requests_delete = self.set_up_patch('searchguard.internalusers.requests.delete')
         self.mocked_requests_delete.return_value = Mock(status_code=200)
-        self.mocked_check_user_exists = self.set_up_patch('searchguard.searchguard.check_user_exists')
+        self.mocked_check_user_exists = self.set_up_patch('searchguard.internalusers.check_user_exists')
         self.mocked_check_user_exists.return_value = True
 
     def test_delete_user_returns_when_successfully_deleted_user(self):
