+++ tests/tests_roles/test_delete_role.py
@@ -3,7 +3,7 @@
 import unittest.mock as mock
 from unittest.mock import Mock
 from tests.helper import BaseTestCase
-from searchguard.searchguard import delete_role
+from searchguard.roles import delete_role
 from searchguard.exceptions import DeleteRoleException
 
 
@@ -12,11 +12,11 @@ class TestDeleteRole(BaseTestCase):
     def setUp(self):
         self.role = "DummyRole"
         self.api_url = "fake_api_url/roles/"
-        self.set_up_patch('searchguard.searchguard.SGAPI', "fake_api_url")
+        self.set_up_patch('searchguard.roles.SGAPI', "fake_api_url")
 
-        self.mocked_requests_delete = self.set_up_patch('searchguard.searchguard.requests.delete')
+        self.mocked_requests_delete = self.set_up_patch('searchguard.roles.requests.delete')
         self.mocked_requests_delete.return_value = Mock(status_code=200)
-        self.mocked_check_role_exists = self.set_up_patch('searchguard.searchguard.check_role_exists')
+        self.mocked_check_role_exists = self.set_up_patch('searchguard.roles.check_role_exists')
         self.mocked_check_role_exists.return_value = True
 
     def test_delete_role_returns_when_successfully_deleted_role(self):
