+++ tests/tests_roles/test_modify_role.py
@@ -4,7 +4,7 @@ import json
 import unittest.mock as mock
 from unittest.mock import Mock
 from tests.helper import BaseTestCase
-from searchguard.searchguard import modify_role
+from searchguard.roles import modify_role
 from searchguard.exceptions import ModifyRoleException
 
 
@@ -14,11 +14,11 @@ class TestModifyRole(BaseTestCase):
         self.role = "DummyRole"
         self.permissions = {"cluster": ["dummyperm"], "indices": {"dummyindice": {"dummytype": ["READ"]}}}
         self.api_url = "fake_api_url/roles/"
-        self.set_up_patch('searchguard.searchguard.SGAPI', "fake_api_url")
+        self.set_up_patch('searchguard.roles.SGAPI', "fake_api_url")
 
-        self.mocked_requests_put = self.set_up_patch('searchguard.searchguard.requests.put')
+        self.mocked_requests_put = self.set_up_patch('searchguard.roles.requests.put')
         self.mocked_requests_put.return_value = Mock(status_code=200)
-        self.mocked_check_role_exists = self.set_up_patch('searchguard.searchguard.check_role_exists')
+        self.mocked_check_role_exists = self.set_up_patch('searchguard.roles.check_role_exists')
         self.mocked_check_role_exists.return_value = True
 
     def test_modify_role_returns_when_successfully_modified_role(self):
